Use createSlice to create action creators and reducers together
This makes it harder to forget to write the reducer after writing the
action creator, or vice versa. It's still necessary to separately define
and export the action creators, but forgetting to do either of those
things will lead to a compile-time error, as long as they're actually
used elsewhere.