Allow adding and editing pipes
This kind of sucks! But there are apparently no better options!

There's a library called graphology which would be lovely to use, but it
gives you an opaque `Graph` object that you can't put in Redux. So we
just store some bonus redundant data to allow O(1) lookups in the ways
we need and hope that everything stays in sync.

Also, the changes to CommandNode avoid crashes when the state and the
props are out of sync with one another ("stale props", as described by
the React documentation).

Currently it's possible to create strange graphs (loops, splits, merges)
-- code will need to be added to forbid these configurations before
there's any hope of producing a plain text version of the graph.