Display the bare bones of the current command in the header
The definition of "current" is a little tricky, I'm sure this isn't the
best that could be done, but it's fine for the time being.

This doesn't currently display anything beyond the basename of the
command (e.g. "cut") -- fixing that will involve properly working out
how the commands for individual nodes should be generated.